+++ InvenTree/common/forms.py
@@ -23,18 +23,3 @@ class SettingEditForm(HelperForm):
         fields = [
             'value'
         ]
-
-
-class UploadFile(forms.Form):
-    ''' Step 1 '''
-    first_name = forms.CharField(max_length=100)
-
-
-class MatchField(forms.Form):
-    ''' Step 2 '''
-    last_name = forms.CharField(max_length=100)
-
-
-class MatchPart(forms.Form):
-    ''' Step 3 '''
-    age = forms.IntegerField()
